refactor(checkpoint): un-split 11 -> 6 files; method-pinning makes subpackages impossible
internal/checkpoint cannot be split into subpackages. 20 of its 52 functions are
methods on *CheckpointManager, spread across six of the ten source files
(alerts, compute, manager, network_helpers, resolution, summary). Go only allows
methods on a type in the package that declares it, so every one of those files is
pinned to whichever package declares CheckpointManager. A "split" would have to
either move CheckpointManager along with all six files — i.e. not a split — or
export the manager's mutex-guarded fields (autoCheckpoint, namedCheckpoints,
namedOrder, pendingAlerts, alertCounter, alertDelivery) so sibling packages could
reach them. That is the internal/queries situation: exporting a lock and its
invariant to satisfy a file count.

So this consolidates instead, following the call graph rather than the filename
prefixes:

  checkpoint_manager.go  = manager + resolution + summary
      The whole GetChangesSince path in one file: resolve the checkpoint,
      compute the diffs, filter by severity, summarize. determineSeverity and
      buildSummary were the last two steps of applySeverityFilter, one call away,
      in a separate file.

  checkpoint_diff.go     = compute + console_helpers + network_helpers + ws_helpers
      The four computeXDiff methods and the classifiers they call. Every symbol
      in the three *_helpers.go files had exactly one caller, and that caller was
      always its computeXDiff in compute.go. The console/network/ws split was one
      algorithm cut into four files.

Filename provenance was not independent evidence here: git log --diff-filter=A
shows all four *_helpers/compute files were minted by a single commit.

Tests move with their code and are renamed to match their new home:
  checkpoint_compute_test.go    -> checkpoint_diff_test.go
  checkpoint_classify_test.go   -> checkpoint_diff_classify_test.go
  checkpoint_resolution_test.go -> checkpoint_manager_test.go

Behavior preservation is mechanical: the exported symbol set (25 symbols) and the
multiset of non-comment, non-import code lines (2,650) are both byte-identical
before and after. Package coverage is unchanged at 98.8%.
