+++ internal/checkpoint/checkpoint_diff_test.go
